Move deluge base entity to separate module
