+++ homeassistant/components/deluge/__init__.py
@@ -17,10 +17,8 @@ from homeassistant.const import (
 )
 from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import ConfigEntryAuthFailed, ConfigEntryNotReady
-from homeassistant.helpers.device_registry import DeviceEntryType, DeviceInfo
-from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
-from .const import CONF_WEB_PORT, DEFAULT_NAME, DOMAIN
+from .const import CONF_WEB_PORT
 from .coordinator import DelugeDataUpdateCoordinator
 
 PLATFORMS = [Platform.SENSOR, Platform.SWITCH]
@@ -61,24 +59,3 @@ async def async_setup_entry(hass: HomeAssistant, entry: DelugeConfigEntry) -> bo
 async def async_unload_entry(hass: HomeAssistant, entry: DelugeConfigEntry) -> bool:
     """Unload a config entry."""
     return await hass.config_entries.async_unload_platforms(entry, PLATFORMS)
-
-
-class DelugeEntity(CoordinatorEntity[DelugeDataUpdateCoordinator]):
-    """Representation of a Deluge entity."""
-
-    _attr_has_entity_name = True
-
-    def __init__(self, coordinator: DelugeDataUpdateCoordinator) -> None:
-        """Initialize a Deluge entity."""
-        super().__init__(coordinator)
-        self._server_unique_id = coordinator.config_entry.entry_id
-        self._attr_device_info = DeviceInfo(
-            configuration_url=(
-                f"http://{coordinator.api.host}:{coordinator.api.web_port}"
-            ),
-            entry_type=DeviceEntryType.SERVICE,
-            identifiers={(DOMAIN, coordinator.config_entry.entry_id)},
-            manufacturer=DEFAULT_NAME,
-            name=DEFAULT_NAME,
-            sw_version=coordinator.api.deluge_version,
-        )
